Move files into the new folder
