+++ src/components/MultiFileUpload/__tests__/MultiFileUploadStatus.test.jsx
@@ -10,14 +10,14 @@ import {
   ENDPOINT_DECLARATION_ATTACHMENTS_PATH,
   ENDPOINT_DECLARATION_PATH,
   ENDPOINT_FILE_UPLOAD_SUPPORTING_DOCUMENTS_PATH,
-} from '../../constants/AppAPIConstants';
+} from '../../../constants/AppAPIConstants';
 import {
   SIGN_IN_URL,
   URL_DECLARATIONID_IDENTIFIER,
   VOYAGE_SUPPORTING_DOCS_UPLOAD_URL,
-} from '../../constants/AppUrlConstants';
+} from '../../../constants/AppUrlConstants';
 import MultiFileUploadForm from '../MultiFileUploadForm';
-import { MAX_SUPPORTING_FILE_SIZE, MAX_SUPPORTING_FILE_SIZE_DISPLAY } from '../../constants/AppConstants';
+import { MAX_SUPPORTING_FILE_SIZE, MAX_SUPPORTING_FILE_SIZE_DISPLAY } from '../../../constants/AppConstants';
 
 const mockedUseNavigate = jest.fn();
 jest.mock('react-router', () => ({
